adiciona lógica de parceiros
