+++ src/app/features/report-generate/report-generate.component.html
@@ -1,26 +1,23 @@
 <form (ngSubmit)="gerarRelatorio()">
-  <!-- Tipo de relatório -->
   <label for="tipo-relatorio">Tipo de Relatório:</label>
   <select id="tipo-relatorio" name="tipoRelatorio" [(ngModel)]="tipoRelatorio" required>
-    <option value="" disabled selected>Selecione um tipo de relatório</option>
+    <option value="" disabled>Selecione um tipo de relatório</option>
     <option value="por-parceiro">Total de doações por parceiro</option>
     <option value="comparativo-parceiro">Comparativo: Doações com e sem parceiro</option>
     <option value="por-mes">Total arrecadado por mês</option>
   </select>
 
   <label for="data-inicio">Data início</label>
-  <input type="date" id="data-inicio" name="dataInicio" [(ngModel)]="dataInicio" required placeholder="Selecione a data de início" />
+  <input type="date" id="data-inicio" name="dataInicio" [(ngModel)]="dataInicio" required />
 
   <label for="data-fim">Data fim</label>
-  <input type="date" id="data-fim" name="dataFim" [(ngModel)]="dataFim" required placeholder="Selecione a data de fim" />
+  <input type="date" id="data-fim" name="dataFim" [(ngModel)]="dataFim" required />
 
-  <label for="tipo-parceiro">Tipo de Parceiro:</label>
+  <label for="tipo-parceiro">Parceiro:</label>
   <select id="parceiro" name="parceiro" [(ngModel)]="parceiro" required>
-    <option value="" disabled selected>Selecione um tipo de parceiro</option>
+    <option value="" disabled>Selecione um parceiro</option>
     <option value="todos">Todos</option>
-    <option value="Parceiro Tech">Parceiro Tech</option>
-    <option value="Parceiro Educa">Parceiro Educa</option>
-    <option value="Parceiro Saúde">Parceiro Saúde</option>
+    <option *ngFor="let p of parceiros" [value]="p.nome">{{ p.nome }}</option>
   </select>
 
   <button type="submit" class="report-btn">Gerar Relatório</button>
